refactor(web): move refresh effects to useRefresh hook
